+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/node/StringConcatenateAssignmentNode.java
@@ -15,7 +15,11 @@ import java.util.Objects;
  * <pre>
  *   <em>variable</em> += <em>expression</em>
  * </pre>
+ *
+ * @deprecated StringConcatenateAssignmentNode is no longer used in CFGs. Instead, an assignment and
+ *     a concatenation node are generated.
  */
+@Deprecated // 2022-03-22
 public class StringConcatenateAssignmentNode extends Node {
     /** The entire tree of the assignment */
     protected final Tree tree;