+++ docs/CHANGELOG.md
@@ -17,6 +17,10 @@ Changed `CFAbstractTransfer.insertIntoStores` from public to protected
 visibility. It is only meant as a utility method for use within a
 transfer function.
 
+Deprecated class `StringConcatenateAssignmentNode` and its usages.
+String concatenate assignments are now desugared to an assignment and a concatenation node instead.
+This avoids error prone duplication of logic.
+
 **Closed issues:**
 typetools#5075.