+++ docs/CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 3.21.3-eisop1 (March ?, 2022)
+Version 3.21.3-eisop1 (March 23, 2022)
 -------------------------------------
 
 **User-visible changes:**
@@ -18,7 +18,8 @@ visibility. It is only meant as a utility method for use within a
 transfer function.
 
 Deprecated class `StringConcatenateAssignmentNode` and its usages.
-String concatenate assignments are now desugared to an assignment and a concatenation node instead.
+String concatenate assignments are now desugared to an assignment and
+a concatenation node instead.
 This avoids error prone duplication of logic.
 
 **Closed issues:**